+++ vs2022/enc_temp_folder/9b985e4736dbab28894c8afa89ac33/TableOfContents.cpp
@@ -1293,7 +1293,7 @@ void CreateToc(MainWindow* win) {
         Edit::CreateArgs eargs;
         eargs.parent = win->hwndTocBox;
         eargs.withBorder = false;
-        eargs.cueText = _TRA("Search Bookmarks");
+        eargs.cueText = _TRN("Search Bookmarks");
         eargs.font = GetDefaultGuiFont(false, false);
         filterEdit->Create(eargs);
     }