+++ vs2022/enc_temp_folder/ec93a6f1c08814e918ca35197e33/SumatraDialogs.cpp
@@ -1571,7 +1571,7 @@ bool Dialog_SetTabColor(HWND hwnd, COLORREF currentColor, bool isUnset, COLORREF
     data.applyToAll = false;
     data.selectedCustomIdx = -1;
     data.previewSelected = true;
-    data.title = _TRA("Set Tab Color");
+    data.title = "Set Tab Color";
     data.showRadioButtons = false;
 
     INT_PTR res = CreateDialogBox(IDD_DIALOG_CHANGE_BG_COLOR, hwnd, Dialog_ChangeBgColor_Proc, (LPARAM)&data);